Refactor library into driver and util files
Refactor library into driver and util files. This library layout helps
organize the gfold API.
